epiccodereview: fix 003-processing sweep2 (10 issues)
frame_stats.rs:
- FS-001: fix luma_variance docstring: range is [0.0, 1.0], not [0, 65025]
- FS-003: alpha_zero + rgb_nonzero: u32 → u64 to prevent overflow on >1.4 GP images;
  propagated to telemetry.rs accumulators (scalar + AVX2 both kernels)
- FS-004: analyze() short-buffer: use px_actual = pixels.len()/4 instead of
  zero_padded() which skewed alpha_zero and luma with phantom pixels; removed
  zero_padded() helper (no longer used)

demosaic.rs:
- DM-001: wire demosaic_rggb_shuffle_simd on wasm32 in demosaic_rggb; eliminates
  3 stack arrays + 24 scalar stores per 8-pixel chunk (3 direct v128 stores instead)
- DM-003: add inline comment on `while col + 8 <= w_max` explaining the intentional
  ≤8-col scalar tail; added to all three wasm32 SIMD loop sites
- DM-006: demosaic_rggb_mhc right-border loop starts at int_end.max(int_start) to
  prevent double-computing col 0 when width==1
- DM-009: extend blue_site_no_pink_cast test to assert shuffle SIMD matches scalar,
  validating the i8x16_shuffle constants on every native CI run
- DM-010: add m12_bayer_mhc_band_grbg_preserves_red_site test: GRBG band path must
  keep direct R sample and match full-frame demosaic_bayer_mhc output
- DM-013: document halo-must-be-even precondition on global_row discard; an odd halo
  would invert local parity vs global frame parity, silently swapping R and B

Skipped (deferred): FS-002/FS-005 (AVX2 hadd/i64 rework, complex), DM-002
(planar_into alloc, needs refactor of do_row), DM-005/DM-007/DM-011/DM-014
(straightforward→complex opportunities), DM-008 (wasm-only micro-opt).

134 tests pass, 8 ignored.

+++ crates/raw-pipeline/src/perceptual/telemetry.rs
@@ -57,7 +57,7 @@ impl Default for RgbHistogram {
 pub fn analyze_fused_scalar(d: &[u8], px: usize) -> TelemetryMetrics {
     let px = px.min(d.len() / 4);
 
-    let (mut a_min, mut a_max, mut a_zero, mut rgb_nz) = (255u32, 0u32, 0u32, 0u32);
+    let (mut a_min, mut a_max, mut a_zero, mut rgb_nz) = (255u32, 0u32, 0u64, 0u64);
     let (mut l_sum, mut l_sq) = (0f64, 0f64);
     let mut lanes = [
         lane_seed(0), lane_seed(1), lane_seed(2), lane_seed(3),
@@ -79,7 +79,7 @@ pub fn analyze_fused_scalar(d: &[u8], px: usize) -> TelemetryMetrics {
         let w = u32::from_le_bytes([d[i], d[i + 1], d[i + 2], d[i + 3]]);
         let lane = p & 7;
         lanes[lane] = (lanes[lane] ^ w).wrapping_mul(PRIME);
-        rgb_nz += (r != 0) as u32 + (g != 0) as u32 + (b != 0) as u32;
+        rgb_nz += (r != 0) as u64 + (g != 0) as u64 + (b != 0) as u64;
         if a < a_min { a_min = a; }
         if a > a_max { a_max = a; }
         if a == 0 { a_zero += 1; }
@@ -117,7 +117,7 @@ unsafe fn analyze_fused_avx2(d: &[u8], px: usize) -> TelemetryMetrics {
     use core::arch::x86_64::*;
 
     let chunks = px / 8;
-    let (mut a_zero, mut rgb_nz) = (0u32, 0u32);
+    let (mut a_zero, mut rgb_nz) = (0u64, 0u64);
     let (mut l_sum, mut l_sq) = (0f64, 0f64);
 
     let mut hv = _mm256_setr_epi32(
@@ -150,8 +150,8 @@ unsafe fn analyze_fused_avx2(d: &[u8], px: usize) -> TelemetryMetrics {
         vmax = _mm256_max_epu8(vmax, _mm256_and_si256(pv, alpha_and));
 
         let zmask = _mm256_movemask_epi8(_mm256_cmpeq_epi8(pv, zero)) as u32;
-        a_zero += (zmask & 0x8888_8888).count_ones();
-        rgb_nz += 24 - (zmask & 0x7777_7777).count_ones();
+        a_zero += (zmask & 0x8888_8888).count_ones() as u64;
+        rgb_nz += 24 - (zmask & 0x7777_7777).count_ones() as u64;
 
         let lo16 = _mm256_cvtepu8_epi16(_mm256_castsi256_si128(pv));
         let hi16 = _mm256_cvtepu8_epi16(_mm256_extracti128_si256(pv, 1));
@@ -210,7 +210,7 @@ unsafe fn analyze_fused_avx2(d: &[u8], px: usize) -> TelemetryMetrics {
         let w = u32::from_le_bytes([d[i], d[i + 1], d[i + 2], d[i + 3]]);
         let lane = p & 7;
         lanes[lane] = (lanes[lane] ^ w).wrapping_mul(PRIME);
-        rgb_nz += (r != 0) as u32 + (g != 0) as u32 + (b != 0) as u32;
+        rgb_nz += (r != 0) as u64 + (g != 0) as u64 + (b != 0) as u64;
         if a < a_min { a_min = a; }
         if a > a_max { a_max = a; }
         if a == 0 { a_zero += 1; }
